Issue #49 - junit integration
Extracted RLOGStatement to own class.
